Remove variables from workflowresultexportservice
