+++ core/gui/src/app/workspace/component/result-exportation/result-exportation.component.ts
@@ -1,10 +1,11 @@
 import { UntilDestroy, untilDestroyed } from "@ngneat/until-destroy";
 import { Component, inject, Input, OnInit } from "@angular/core";
-import { environment } from "../../../../environments/environment";
 import { WorkflowResultExportService } from "../../service/workflow-result-export/workflow-result-export.service";
 import { DashboardDataset } from "../../../dashboard/type/dashboard-dataset.interface";
 import { DatasetService } from "../../../dashboard/service/user/dataset/dataset.service";
 import { NZ_MODAL_DATA, NzModalRef } from "ng-zorro-antd/modal";
+import { WorkflowActionService } from "../../service/workflow-graph/model/workflow-action.service";
+import { WorkflowResultService } from "../../service/workflow-result/workflow-result.service";
 
 @UntilDestroy()
 @Component({
@@ -35,7 +36,9 @@ export class ResultExportationComponent implements OnInit {
   constructor(
     public workflowResultExportService: WorkflowResultExportService,
     private modalRef: NzModalRef,
-    private datasetService: DatasetService
+    private datasetService: DatasetService,
+    private workflowActionService: WorkflowActionService,
+    private workflowResultService: WorkflowResultService
   ) {}
 
   ngOnInit(): void {
@@ -51,10 +54,19 @@ export class ResultExportationComponent implements OnInit {
   }
 
   updateOutputType(): void {
-    this.workflowResultExportService.determineOutputTypeForHighlightedOperator();
-    this.isTableOutput = this.workflowResultExportService.isTableOutput;
-    this.isVisualizationOutput = this.workflowResultExportService.isVisualizationOutput;
-    this.containsBinaryData = this.workflowResultExportService.containsBinaryData;
+    const highlightedOperatorIds = this.workflowActionService.getJointGraphWrapper().getCurrentHighlightedOperatorIDs();
+    if (highlightedOperatorIds.length === 1) {
+      const operatorId = highlightedOperatorIds[0];
+      const outputTypes = this.workflowResultService.determineOutputTypes(operatorId);
+      this.isTableOutput = outputTypes.isTableOutput;
+      this.isVisualizationOutput = outputTypes.isVisualizationOutput;
+      this.containsBinaryData = outputTypes.containsBinaryData;
+    } else {
+      // TODO: handle multiple operators
+      this.isTableOutput = false;
+      this.isVisualizationOutput = false;
+      this.containsBinaryData = false;
+    }
   }
 
   onUserInputDatasetName(event: Event): void {
